Move bot implementations out of player/ into a dedicated ai/ package
PassBot/GreedyBot/CappedGreedyBot were sitting under
gamecore/player/ alongside CLIPlayer/NetworkPlayer, but they aren't a
player-seat abstraction -- they're bot decision logic that happens to
implement BotInterface. Move them to highsociety/code/ai/ (an
already-scaffolded, previously-empty package) and their tests to
tests/ai/ to match. No behavior change, just import paths.
