+++ BOT_API.md
@@ -7,8 +7,10 @@ Two ways to build a bot, depending on where it runs:
   `get_bid`, `choose_painting_to_discard`, `send_message` — then hand an instance to `PlayGame`
   directly. `BasePlayer` gives you bidding/card bookkeeping for free; `BotInterface` is an ABC, so
   forgetting one of the three raises `TypeError` immediately at construction, not partway through a
-  game. `CLIPlayer`/`NetworkPlayer` are the two existing implementations — read either as a
-  reference. Simplest option if your bot lives in this codebase.
+  game. `CLIPlayer`/`NetworkPlayer` (`player.py`'s human-facing implementations) work as a
+  reference, but `highsociety/code/ai/` (`pass_bot.py`, `greedy_bot.py`, `capped_greedy_bot.py`)
+  are actual bots to read instead — see `dev_tools/simulate_bots.py` to watch them play live.
+  Simplest option if your bot lives in this codebase.
 - **Remote** — connect over the network like any other player (`network_client.py`'s protocol),
   in any language. This is what you want for an external bot-building competition, since entrants
   don't need this repo at all — just a TCP socket and JSON.
