+++ highsociety/code/ai/greedy_bot.py
