+++ highsociety/code/gamecore/dev_tools/simulate_bots.py
@@ -15,9 +15,9 @@ Usage:
 import argparse
 
 from highsociety.code.gamecore.game_manager.gameplay import PlayGame
-from highsociety.code.gamecore.player.capped_greedy_bot import CappedGreedyBot
-from highsociety.code.gamecore.player.greedy_bot import GreedyBot
-from highsociety.code.gamecore.player.pass_bot import PassBot
+from highsociety.code.ai.capped_greedy_bot import CappedGreedyBot
+from highsociety.code.ai.greedy_bot import GreedyBot
+from highsociety.code.ai.pass_bot import PassBot
 
 _BOT_TYPES = {
     "pass": PassBot,
