+++ tests/ai/test_capped_greedy_bot.py
@@ -1,8 +1,8 @@
 import pytest
 
-from highsociety.code.gamecore.player.capped_greedy_bot import CappedGreedyBot
-from highsociety.code.gamecore.player.greedy_bot import GreedyBot
-from highsociety.code.gamecore.player.pass_bot import PassBot
+from highsociety.code.ai.capped_greedy_bot import CappedGreedyBot
+from highsociety.code.ai.greedy_bot import GreedyBot
+from highsociety.code.ai.pass_bot import PassBot
 from highsociety.code.gamecore.game_manager.gameplay import PlayGame
 from highsociety.code.gamecore.components_module.painting import Painting
